// sdk/typescript/src/generated/arka.ts
export interface Asset {
  contract: string;
}
// ...
export interface SwapStep {
  adapter: string;
  amount_in: i128;
  asset_in: Asset;
  asset_out: Asset;
  min_out: i128;
  pool_id: u128;
  router_addr: string;
}
// ...
export interface FeeStructure {
  deposit_bps: i32;
  mgmt_bps: i32;
  perf_bps: i32;
  redeem_bps: i32;
}


export interface BlendPosition {
  asset: string;
  collateral_amount: i128;
  debt_amount: i128;
  market_id: u128;
}


export interface CreditPosition {
  asset: string;
  collateral_amount: i128;
  debt_amount: i128;
  market_id: u128;
}

export type CreditProtocol = {tag: "Blend", values: void};
// ...
export interface BlendMarketStatus {
  debt_value: i128;
  has_disabled_reserve: boolean;
  has_future_oracle_timestamp: boolean;
  has_invalid_oracle_data: boolean;
  has_live_pricing: boolean;
  has_stale_oracle: boolean;
  health_factor: i128;
  market_id: u128;
  max_oracle_age: u64;
  min_health_factor: i128;
  nav_blocked: boolean;
  oracle_age: u64;
  pool_status: u32;
  risky_actions_blocked: boolean;
}
// ...
export interface CreditMarketStatus {
  debt_value: i128;
  has_disabled_reserve: boolean;
  has_future_oracle_timestamp: boolean;
  has_invalid_oracle_data: boolean;
  has_live_pricing: boolean;
  has_stale_oracle: boolean;
  health_factor: i128;
  market_id: u128;
  max_oracle_age: u64;
  min_health_factor: i128;
  nav_blocked: boolean;
  oracle_age: u64;
  pool_status: u32;
  risky_actions_blocked: boolean;
}